Converted to Room, code compiles.

+++ app/src/main/java/org/willemsens/player/view/main/music/songs/SongsFragment.java
@@ -128,13 +128,13 @@ public class SongsFragment extends Fragment implements PopupMenu.OnMenuItemClick
             final MenuItem menuItemAll = albumsMenu.getItem(i++);
             menuItemAll.setCheckable(true);
             menuItemAll.setChecked(true);
-            for (Iterator<Map.Entry<Integer, Boolean>> it = filter.getAlbumIterator(); it.hasNext();) {
-                final Map.Entry<Integer, Boolean> entry = it.next();
-                final Integer albumId = entry.getKey();
+            for (Iterator<Map.Entry<Long, Boolean>> it = filter.getAlbumIterator(); it.hasNext();) {
+                final Map.Entry<Long, Boolean> entry = it.next();
+                final Long albumId = entry.getKey();
                 // TODO: performance: fetch all artists at once!
                 // TODO: make it sorted: fetch all artists at once!
                 final Album album = musicDao.findAlbum(albumId);
-                albumsMenu.add(NONE, album.id, NONE, album.name);
+                albumsMenu.add(NONE, (int)album.id, NONE, album.name);
                 final MenuItem menuItem = albumsMenu.getItem(i++);
                 menuItem.setCheckable(true);
                 if (entry.getValue()) {
@@ -149,14 +149,14 @@ public class SongsFragment extends Fragment implements PopupMenu.OnMenuItemClick
             final MenuItem menuItemAll = artistsMenu.getItem(i++);
             menuItemAll.setCheckable(true);
             menuItemAll.setChecked(true);
-            for (Iterator<Map.Entry<Integer, Boolean>> it = filter.getArtistIterator(); it.hasNext();) {
-                final Map.Entry<Integer, Boolean> entry = it.next();
-                final Integer artistId = entry.getKey();
+            for (Iterator<Map.Entry<Long, Boolean>> it = filter.getArtistIterator(); it.hasNext();) {
+                final Map.Entry<Long, Boolean> entry = it.next();
+                final Long artistId = entry.getKey();
                 // TODO: performance: fetch all artists at once!
                 // TODO: make it sorted: fetch all artists at once!
                 final Artist artist = musicDao.findArtist(artistId);
                 // Negative ID means it's an artist
-                artistsMenu.add(NONE, -artist.id, NONE, artist.name);
+                artistsMenu.add(NONE, (int)-artist.id, NONE, artist.name);
                 final MenuItem menuItem = artistsMenu.getItem(i++);
                 menuItem.setCheckable(true);
                 if (entry.getValue()) {
